back-end/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
autopep8==1.5.5
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.6
django-cors-headers==3.7.0
django-debug-toolbar==3.2
django-rest-swagger==2.2.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
idna==2.10
imap-tools==0.38.1
IMAPClient==2.2.0
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mypy-extensions==0.4.3
openapi-codec==1.3.2
pathspec==0.8.1
psycopg2==2.8.6
pycodestyle==2.6.0
PyJWT==2.0.1
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
